UINT_MAX -> ULONG_MAX to avoid truncating negative 64-bit values into positive 32-bit ones
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6136 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/lib_errors.h
@@ -252,9 +252,9 @@ STMT(\
 	i64 err64__ = (i64)(expression);\
 	if(err64__ < 0)\
 	{\
-		LibError err__ = (LibError)(err64__ & UINT_MAX);\
+		LibError err__ = (LibError)(err64__ & ULONG_MAX);\
 		DEBUG_WARN_ERR(err__);\
-		return (LibError)(err__ & UINT_MAX);\
+		return (LibError)(err__ & ULONG_MAX);\
 	}\
 )
 
@@ -265,7 +265,7 @@ STMT(\
 	i64 err64__ = (i64)(expression);\
 	if(err64__ < 0)\
 	{\
-		LibError err__ = (LibError)(err64__ & UINT_MAX);\
+		LibError err__ = (LibError)(err64__ & ULONG_MAX);\
 		return err__;\
 	}\
 )
@@ -284,7 +284,7 @@ STMT(\
 	i64 err64__ = (i64)(expression);\
 	if(err64__ < 0)\
 	{\
-		LibError err__ = (LibError)(err64__ & UINT_MAX);\
+		LibError err__ = (LibError)(err64__ & ULONG_MAX);\
 		DEBUG_WARN_ERR(err__);\
 		throw err__;\
 	}\
@@ -297,7 +297,7 @@ STMT(\
 	i64 err64__ = (i64)(expression);\
 	if(err64__ < 0)\
 	{\
-		LibError err__ = (LibError)(err64__ & UINT_MAX);\
+		LibError err__ = (LibError)(err64__ & ULONG_MAX);\
 		DEBUG_WARN_ERR(err__);\
 		return;\
 	}\
@@ -310,7 +310,7 @@ STMT(\
 	i64 err64__ = (i64)(expression);\
 	if(err64__ < 0)\
 	{\
-		LibError err__ = (LibError)(err64__ & UINT_MAX);\
+		LibError err__ = (LibError)(err64__ & ULONG_MAX);\
 		DEBUG_WARN_ERR(err__);\
 	}\
 )
